Pull reposition related changes to a new VM
